Deprecate VSCImportTest and move it to a dedicated module (the prospector imports were introducing sideeffects)
