+++ lib/vsc/install/shared_setup.py
@@ -146,7 +146,7 @@ URL_GHUGENT_HPCUGENT = 'https://github.ugent.be/hpcugent/%(name)s'
 
 RELOAD_VSC_MODS = False
 
-VERSION = '0.9.15'
+VERSION = '0.9.16'
 
 log.info('This is (based on) vsc.install.shared_setup %s' % VERSION)
