+++ test/00-import.py
@@ -1,5 +1,5 @@
 # don't import VSCImportTest, it will trigger the tests
-import vsc.install.testing
+import vsc.install.commontest
 
-class ImportTest(vsc.install.testing.VSCImportTest):
+class ImportTest(vsc.install.commontest.CommonTest):
     EXCLUDE_MODS = ['^vsc\.fancylogger$'] # it requires vsc-base
